+++ R/ie.R
@@ -15,12 +15,13 @@
 #'  \item{IETEST}{Inclusion/Exclusion Criterion Test Name}
 #'  \item{IETESTCD}{Inclusion/Exclusion Criterion Short Name}
 #'  \item{IEORRES}{Result or Finding in Original Units}
-#'  \item{IESTRESC}{Character Result/Finding in Standard Format}
+#'  \item{IESTRESC}{Character Result Standard Format}
 #'  \item{IEDTC}{Date/Time of Collection}
 #' }
-#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{suppho}} \code{\link{ts}}
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{pr}} \code{\link{suppho}} \code{\link{supppc}} \code{\link{supppr}} \code{\link{ts}}
 #' @keywords datasets ie
 #' @name ie
 #' @examples
-#' head(data("ie"))
+#'  head(data("ie"))
 "ie"
+