+++ R/suppho.R
@@ -13,9 +13,10 @@
 #'  \item{QLABEL}{Qualifier Variable Label}
 #'  \item{QVAL}{Qualifier Value}
 #' }
-#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{pr}} \code{\link{suppho}} \code{\link{ts}}
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{pr}} \code{\link{suppho}} \code{\link{supppc}} \code{\link{supppr}} \code{\link{ts}}
 #' @keywords datasets suppho
 #' @name suppho
 #' @examples
-#' head(data("suppho"))
+#'  head(data("suppho"))
 "suppho"
+