+++ NEWS.md
@@ -14,6 +14,7 @@
 * Added variables to MH:
   * MHSTDY
   * MHENDY
+  * Updated MH for MHCAT values
 
 * Added